Rename to cloudfoundry-resources
Uses plural form to align with aws-resources and gcp-resources subprojects.
