--- cloudfoundry-resource/src/main/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.ResourceProvider
@@ -1,1 +1,0 @@
-io.opentelemetry.contrib.cloudfoundry.resource.CloudFoundryResourceProvider
