+++ cloudfoundry-resources/src/test/java/io/opentelemetry/contrib/cloudfoundry/resources/CloudFoundryResourceTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.cloudfoundry.resource;
+package io.opentelemetry.contrib.cloudfoundry.resources;
 
 import static org.assertj.core.api.Assertions.assertThat;
