+++ cloudfoundry-resources/src/test/resources/vcap_application.json
