+++ settings.gradle.kts
@@ -33,7 +33,7 @@ include(":aws-xray")
 include(":aws-xray-propagator")
 include(":baggage-processor")
 include(":compressors:compressor-zstd")
-include("cloudfoundry-resource")
+include("cloudfoundry-resources")
 include(":consistent-sampling")
 include(":dependencyManagement")
 include(":disk-buffering")
